refactor: collapse batch module into hl7v2
