+++ crates/hl7v2-batch/Cargo.toml
@@ -13,9 +13,9 @@ keywords = ["hl7", "healthcare", "batch", "file", "processing"]
 categories = ["parser-implementations"]
 
 [dependencies]
-thiserror = { workspace = true }
-hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
-hl7v2-parser = { version = "1.2.0", path = "../hl7v2-parser" }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "batch",
+] }
 
 [dev-dependencies]
 hl7v2-test-utils = { path = "../hl7v2-test-utils" }
