+++ crates/hl7v2-batch/tests/integration_tests.rs
@@ -3,8 +3,8 @@
 //! These tests verify batch parsing and creation works correctly with
 //! real-world HL7 batch scenarios.
 
+use hl7v2::parse;
 use hl7v2_batch::{Batch, BatchType, FileBatch, parse_batch};
-use hl7v2_parser::parse;
 
 // ============================================================================
 // Real-World Batch File Tests
